+++ src/main/java/org/akip/domain/ProcessMember.java
@@ -1,0 +1,95 @@
+package org.akip.domain;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import org.hibernate.annotations.BatchSize;
+import org.hibernate.annotations.Cache;
+import org.hibernate.annotations.CacheConcurrencyStrategy;
+
+import javax.persistence.*;
+import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+
+@Entity
+@Table(name = "process_member")
+@Cache(usage = CacheConcurrencyStrategy.READ_WRITE)
+public class ProcessMember implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
+    @SequenceGenerator(name = "sequenceGenerator", sequenceName = "akip_hibernate_sequence")
+    private Long id;
+
+    @Column
+    private String username;
+
+    @ManyToOne
+    private ProcessDefinition processDefinition;
+
+    @JsonIgnore
+    @ManyToMany
+    @JoinTable(
+            name = "process_member_process_role",
+            joinColumns = { @JoinColumn(name = "process_member_id", referencedColumnName = "id") },
+            inverseJoinColumns = { @JoinColumn(name = "process_role_id", referencedColumnName = "id") }
+    )
+    @Cache(usage = CacheConcurrencyStrategy.NONSTRICT_READ_WRITE)
+    @BatchSize(size = 20)
+    private List<ProcessRole> processRoles = new ArrayList<>();
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public ProcessDefinition getProcessDefinition() {
+        return processDefinition;
+    }
+
+    public void setProcessDefinition(ProcessDefinition processDefinition) {
+        this.processDefinition = processDefinition;
+    }
+
+    public List<ProcessRole> getProcessRoles() {
+        return processRoles;
+    }
+
+    public void setProcessRoles(List<ProcessRole> processRoles) {
+        this.processRoles = processRoles;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        ProcessMember that = (ProcessMember) o;
+        return Objects.equals(id, that.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
+    }
+
+    @Override
+    public String toString() {
+        return "ProcessMember{" +
+                "id=" + id +
+                ", processDefinition=" + processDefinition +
+                '}';
+    }
+}