+++ src/main/java/org/akip/repository/TenantMemberRepository.java
@@ -17,7 +17,17 @@ public interface TenantMemberRepository extends JpaRepository<TenantMember, Long
 
     List<TenantMember> findByTenantId(Long tenantId);
 
+    List<TenantMember> findTenantMembersByUsername(String username);
+
     @Query("select tenant from TenantMember where username = ?1")
     List<Tenant> findTenantByTenantMemberUsername(String username);
+
+    @Query("SELECT CONCAT(tm.tenant.identifier, '.', tr.name) FROM TenantMember tm JOIN tm.tenantRoles tr WHERE tm.username = ?1")
+    List<String> findTenantRolesWithDefinitionIdPrefixByUsername(String username);
+
+    @Query("SELECT CONCAT(tm.tenant.identifier, '.*') FROM TenantMember tm WHERE tm.username = ?1")
+    List<String> findTenantWildcardPrefixByUsername (String username);
+
+    List<TenantMember> findTenantMembersByTenantId(Long id);
 
 }