+++ src/main/java/org/akip/service/ProcessInstanceService.java
@@ -129,7 +129,7 @@ public class ProcessInstanceService {
 
         processInstance.setCamundaProcessInstanceId(camundaProcessInstance.getProcessInstanceId());
         ProcessInstanceDTO processInstanceSaved = processInstanceMapper.toDto(processInstanceRepository.save(processInstance));
-        synchronizeAttachmentsAndNotesAndUpdateTemporaryProcessInstance(processInstanceDTO, processInstance);
+        synchronizeAttachmentsAndNotesAndUpdateTemporaryProcessInstance(processInstanceDTO.getTemporaryProcessInstance(), processInstanceSaved);
         runtimeService.setVariable(camundaProcessInstance.getProcessInstanceId(), CamundaConstants.PROCESS_INSTANCE, processInstanceSaved);
         return processInstanceSaved;
     }
@@ -161,10 +161,9 @@ public class ProcessInstanceService {
                 .setVariables(params)
                 .execute();
 
-
         processInstance.setCamundaProcessInstanceId(camundaProcessInstance.getProcessInstanceId());
         ProcessInstanceDTO processInstanceSaved = processInstanceMapper.toDto(processInstanceRepository.save(processInstance));
-        synchronizeAttachmentsAndNotesAndUpdateTemporaryProcessInstance(processInstanceDTO, processInstance);
+        synchronizeAttachmentsAndNotesAndUpdateTemporaryProcessInstance(processInstanceDTO.getTemporaryProcessInstance(), processInstanceSaved);
         runtimeService.setVariable(camundaProcessInstance.getProcessInstanceId(), CamundaConstants.PROCESS_INSTANCE, processInstanceSaved);
         return processInstanceSaved;
     }
@@ -187,9 +186,13 @@ public class ProcessInstanceService {
     private ProcessInstance createWithoutTenant(String bpmnProcessDefinitionId, String businessKey, IProcessEntity processEntity) {
         log.debug("Request to create a processInstance by bpmnProcessDefinitionId: {}", bpmnProcessDefinitionId);
 
+        //TODO: These methods (create) should be refactored: 1. different return types, handling domain and DTO,
+        TemporaryProcessInstanceDTO temporaryProcessInstance = processEntity.getProcessInstance().getTemporaryProcessInstance();
+
         ProcessDefinition processDefinition = processDefinitionRepository
                 .findByBpmnProcessDefinitionId(bpmnProcessDefinitionId)
                 .orElseThrow();
+
         ProcessDeployment processDeployment = processDeploymentRepository
                 .findByProcessDefinitionIdAndStatusIsActiveAndTenantIsNull(processDefinition.getId())
                 .orElseThrow();
@@ -216,7 +219,11 @@ public class ProcessInstanceService {
                 .execute();
 
         processInstance.setCamundaProcessInstanceId(camundaProcessInstance.getProcessInstanceId());
-        return processInstanceRepository.save(processInstance);
+
+        ProcessInstance processInstanceSaved = processInstanceRepository.save(processInstance);
+        synchronizeAttachmentsAndNotesAndUpdateTemporaryProcessInstance(temporaryProcessInstance, processInstanceMapper.toDto(processInstanceSaved));
+
+        return processInstanceSaved;
     }
 
     private ProcessInstance createWithTenant(
@@ -260,7 +267,11 @@ public class ProcessInstanceService {
                 .execute();
 
         processInstance.setCamundaProcessInstanceId(camundaProcessInstance.getProcessInstanceId());
-        return processInstanceRepository.save(processInstance);
+
+        ProcessInstance processInstanceSaved = processInstanceRepository.save(processInstance);
+        synchronizeAttachmentsAndNotesAndUpdateTemporaryProcessInstance(processEntity.getProcessInstance().getTemporaryProcessInstance(), processInstanceMapper.toDto(processInstanceSaved));
+
+        return processInstanceSaved;
     }
 
     /**
@@ -316,12 +327,12 @@ public class ProcessInstanceService {
     }
 
     private void synchronizeAttachments(
-            TemporaryProcessInstanceDTO temporaryProcessInstance,
-            ProcessInstance processInstance
+            Long temporaryProcessInstanceId,
+            Long processInstanceId
     ) {
         List<Attachment> attachments = attachmentRepository.findByEntityNameAndEntityId(
                 TemporaryProcessInstance.class.getSimpleName(),
-                temporaryProcessInstance.getId()
+                temporaryProcessInstanceId
         );
         if (attachments.isEmpty()) {
             return;
@@ -332,19 +343,19 @@ public class ProcessInstanceService {
                     AttachmentEntity attachmentEntityProcessInstance = new AttachmentEntity();
                     attachmentEntityProcessInstance.setAttachment(attachment);
                     attachmentEntityProcessInstance.setEntityName(ProcessInstance.class.getSimpleName());
-                    attachmentEntityProcessInstance.setEntityId(processInstance.getId());
+                    attachmentEntityProcessInstance.setEntityId(processInstanceId);
                     attachmentEntityRepository.save(attachmentEntityProcessInstance);
                 }
         );
     }
 
     private void synchronizeNotes(
-            TemporaryProcessInstanceDTO temporaryProcessInstance,
-            ProcessInstance processInstance
+            Long temporaryProcessInstanceId,
+            Long processInstanceId
     ) {
         List<Note> notes = noteRepository.findByEntityNameAndEntityId(
                 TemporaryProcessInstance.class.getSimpleName(),
-                temporaryProcessInstance.getId()
+                temporaryProcessInstanceId
         );
         if (notes.isEmpty()) {
             return;
@@ -354,44 +365,30 @@ public class ProcessInstanceService {
                     NoteEntity noteEntityProcessInstance = new NoteEntity();
                     noteEntityProcessInstance.setNote(note);
                     noteEntityProcessInstance.setEntityName(ProcessInstance.class.getSimpleName());
-                    noteEntityProcessInstance.setEntityId(processInstance.getId());
+                    noteEntityProcessInstance.setEntityId(processInstanceId);
                     noteEntityRepository.save(noteEntityProcessInstance);
+                }
+        );
+
+        //Closing Notes
+        notes.forEach(
+                note -> {
+                    note.setStatus("CLOSED");
+                    noteRepository.save(note);
                 }
         );
     }
 
-    private void synchronizeAttachmentsAndNotesAndUpdateTemporaryProcessInstance(ProcessInstanceDTO processInstanceDTO, ProcessInstance processInstance) {
-        if (processInstanceDTO.getTemporaryProcessInstance() == null) {
+    private void synchronizeAttachmentsAndNotesAndUpdateTemporaryProcessInstance(TemporaryProcessInstanceDTO temporaryProcessInstanceDTO, ProcessInstanceDTO processInstanceSaved) {
+        if (temporaryProcessInstanceDTO == null) {
             return;
         }
-        synchronizeAttachments(processInstanceDTO.getTemporaryProcessInstance(), processInstance);
-        synchronizeNotes(processInstanceDTO.getTemporaryProcessInstance(), processInstance);
-        temporaryProcessInstanceRepository.updateProcessInstanceIdById(processInstance, processInstanceDTO.getTemporaryProcessInstance().getId());
+        synchronizeAttachments(temporaryProcessInstanceDTO.getId(), processInstanceSaved.getId());
+        synchronizeNotes(temporaryProcessInstanceDTO.getId(), processInstanceSaved.getId());
+        TemporaryProcessInstance temporaryProcessInstance = temporaryProcessInstanceRepository.findById(temporaryProcessInstanceDTO.getId()).get();
+        temporaryProcessInstance.setProcessInstance(new ProcessInstance());
+        temporaryProcessInstance.getProcessInstance().setId(processInstanceSaved.getId());
+        temporaryProcessInstanceRepository.save(temporaryProcessInstance);
     }
-
-//    TODO: This method should implement any type of pagination.
-//          Otherwise it may retrieve a huge amount of data
-//    public List<ProcessInstanceDTO> findByProcessDefinition(String idOrBpmnProcessDefinitionId) {
-//        ProcessDefinitionDTO processDefinitionDTO = processDefinitionService
-//            .findByIdOrBpmnProcessDefinitionId(idOrBpmnProcessDefinitionId)
-//            .orElseThrow();
-//        return processInstanceRepository
-//            .findByProcessDefinitionId(processDefinitionDTO.getId())
-//            .stream()
-//            .map(processInstanceMapper::toDto)
-//            .collect(Collectors.toList());
-//    }
-
-    //    TODO: This method should implement any type of pagination.
-    //          Otherwise it may retrieve a huge amount of data
-//    @Transactional(readOnly = true)
-//    public List<ProcessInstanceDTO> findAll() {
-//        log.debug("Request to get all ProcessInstances");
-//        return processInstanceRepository
-//            .findAll()
-//            .stream()
-//            .map(processInstanceMapper::toDto)
-//            .collect(Collectors.toCollection(LinkedList::new));
-//    }
 
 }