+++ src/main/java/org/akip/service/dto/TaskInstanceDTO.java
@@ -52,6 +52,8 @@ public class TaskInstanceDTO implements Serializable {
 
     private List<String> candidateGroups = new ArrayList<>();
 
+    private List<String> computedCandidateGroups = new ArrayList<>();
+
     private String connectorName;
 
     private String connectorConfigName;
@@ -204,6 +206,14 @@ public class TaskInstanceDTO implements Serializable {
 
     public void setCandidateGroups(List<String> candidateGroups) {
         this.candidateGroups = candidateGroups;
+    }
+
+    public List<String> getComputedCandidateGroups() {
+        return computedCandidateGroups;
+    }
+
+    public void setComputedCandidateGroups(List<String> computedCandidateGroups) {
+        this.computedCandidateGroups = computedCandidateGroups;
     }
 
     public String getConnectorName() {